pyproject.toml
[tool.poetry]
name = "lacrema-fastapi"
version = "1.0.0"
description = "API for LaCrema restaurant"
authors = ["Andriy Stefanyuk <stefanyuk.andriyigor@gmail.com>"]
packages = [
  { include = "restaurantservice" },
]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pydantic = "^1.9.0"
poethepoet = "^0.13.1"
asyncpg = "^0.25.0"
SQLAlchemy = "^1.4.36"
requests = "^2.27.1"
httpx = "^0.22.0"
python-jose = "^3.3.0"


[tool.poetry.dev-dependencies]
pylint = "^2.13.7"
pytest = "^7.1.2"
isort = "^5.10.1"
black = "^22.3.0"
poethepoet = "^0.13.1"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.3"
requests = "^2.27.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
test.cmd = "pytest --cov=updateservice tests"
test.env = {LA_CREMA_DB_CONNECTION_STRING = "postgresql+asyncpg://lacrema:lacrema123@lacrema_db:5432/lacrema_db"}

api_test.cmd = "pytest api_tests/"
api_test.env = {LA_CREMA_DB_CONNECTION_STRING = "postgresql+asyncpg://lacrema:lacrema123@localhost:5432/lacrema_db"}

start.cmd = "docker-compose up"
stop.cmd = "docker-compose down"

lint = "pylint updateservice tests api_tests"


[tool.isort]
profile = "black"
src_paths = ["restaurantservice", "tests", "api_tests"]
